+++ karmaConf.js
@@ -14,6 +14,16 @@ module.exports = function(config) {
       'templates/*.html': 'ng-html2js'
     },
 
+    // HTML templates mock
+    ngHtml2JsPreprocessor: {
+      moduleName: 'templates',
+      cacheIdFromPath: function(filepath) {
+        return filepath.replace(/^(.*\/)(.*)$/, function(match, match1, match2) {
+          return 'opl-' + match2;
+        });
+      }
+    },
+
     // List of files / patterns to load in the browser
     files: [
       'node_modules/angular/angular.js',
@@ -23,7 +33,6 @@ module.exports = function(config) {
       'node_modules/angular-mocks/angular-mocks.js',
       'node_modules/video.js/dist/video.js',
       'templates/*.html',
-      'tests/init.js',
       'js/PlayerApp.js',
       'js/**/*.js',
       'tests/*.js'