+++ tasks/concat.js
@@ -9,8 +9,13 @@ module.exports = {
   },
   js: {
 
-    // Match all JavaScript files to concat (PlayerApp.js must be the first one)
-    src: ['<%= player.jsPath %>/PlayerApp.js', '<%= player.jsPath %>/**/*.js'],
+    // Match all JavaScript files to concat
+    // grunt ngtemplates:player needs to be executed before this task
+    src: [
+      '<%= player.jsPath %>/PlayerApp.js',
+      '<%= player.buildPath %>/openveo-player.templates.js',
+      '<%= player.jsPath %>/**/*.js'
+    ],
 
     // Destination file
     dest: '<%= player.distPath %>/openveo-player.js'