+++ tasks/player.js
@@ -2,6 +2,8 @@
 
 module.exports = {
   basePath: ['.'],
+  buildPath: '<%= player.basePath %>/build',
+  templatesPath: '<%= player.basePath %>/templates',
   jsPath: ['<%= player.basePath %>/js/'],
   distPath: ['<%= player.basePath %>/dist'],
   sass: ['<%= player.basePath %>/css']